+++ kernel/irq/timer.h
@@ -1,7 +1,7 @@
 #ifndef IRQ_TIMER_H
 #define IRQ_TIMER_H
 
-extern unsigned int sys_ticks;
+extern unsigned long sys_ticks;
 
 void timer_init();
